+++ PiOS/lib/libpi/include-common/uart.h
@@ -1,12 +1,22 @@
+/**
+ * Jonathan Kula | CS140E | Winter Quarter 2019
+ * File: uart.h
+ * --------------
+ * Provides an interface for the Raspberry Pi's miniUART.
+ */
+
 #ifndef UART_DEFS
 #define UART_DEFS
 
 #include <integer.h>
 
+/** Sets up the miniUART for use. */
 void uart_init(void);
 
-int uart_getc(void);
+/** Gets a character from the peer, blocking until something is recieved. */
+uint8 uart_getc(void);
 
-void uart_putc(uint32 c);
+/** Sends a character to the peer, blockiing until it is queued for sending. */
+void uart_putc(uint8 c);
 
 #endif
